team memberhsip style changes
from design and user research feedback.
