+++ app/views/people/_form.html.haml
@@ -33,7 +33,7 @@
           = render 'membership_fields', membership_f: membership_f, org_structure: @org_structure, person: @person
 
       = link_to 'Join another team', add_membership_people_path(id: @person.to_param),
-        id: 'add_membership', class: 'button-secondary'
+        id: 'add_membership'
 
   %h3.heading-medium.with-border-lg
     = "Current Projects"
