+++ app/views/shared/_breadcrumb.html.haml
@@ -1,6 +1,5 @@
-%li{ class: "breadcrumb-#{breadcrumb_counter}" }<
-  - if breadcrumb.is_a?(String) || !(show_links)
-    = breadcrumb
-  - else
-    = link_to_breadcrumb_name_unless_current breadcrumb, breadcrumb_counter,
-      starts_with_home: starts_with_home, starts_with_root_team: starts_with_root_team
+= crumb_to(breadcrumb,
+           {  index: breadcrumb_counter,
+              starts_with_home: starts_with_home,
+              starts_with_root_team: starts_with_root_team,
+              show_links: show_links } )
