+++ app/views/shared/_breadcrumbs.html.haml
@@ -1,6 +1,6 @@
 - unless items.empty?
   .breadcrumbs
     %ol= render partial: "shared/breadcrumb", collection: items,
-      locals: { show_links: show_links,
+      locals: {show_links: show_links,
         starts_with_home: starts_with_home,
-        starts_with_root_team: starts_with_root_team }
+        starts_with_root_team: starts_with_root_team}
